# agent-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/accounts": {
      "post": {
        "operationId": "AccountController_addOrUpdateAccount",
        "summary": "Create or update an account",
        "description": "Creates a new account or updates an existing one based on the input provided.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrUpdateAccountDto"
              },
              "example": {
                "id": "e1104336-d582-4779-b75b-64deb53e9f6c",
                "name": "Example Corp"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account successfully created or updated.",
            "content": {
              "application/json": {
                "example": {
                  "status": "SUCCESS",
                  "data": {
                    "id": "e1104336-d582-4779-b75b-64deb53e9f6c"
                  },
                  "timestamp": "2025-07-25T08:47:11.993Z"
                }
              }
            }
          },
          "400": {
            "description": "Account creation or update failed.",
            "content": {
              "application/json": {
                "example": {
                  "status": "FAILURE",
                  "message": "",
                  "errorCode": "FAILED_ACCOUNT_UPSERTION",
                  "timestamp": "2025-07-25T08:48:22.634Z",
                  "path": "/agent-backend/accounts"
                }
              }
            }
          }
        },
        "tags": ["Account"]
      }
    },
    "/accounts/{id}": {
      "get": {
        "operationId": "AccountController_getAccount",
        "summary": "Get account by ID",
        "description": "Retrieves the account details for the provided account ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the account to retrieve.",
            "schema": {
              "type": "string",
              "example": "e1104336-d582-4779-b75b-64deb53e9f6c"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account successfully retrieved.",
            "content": {
              "application/json": {
                "example": {
                  "status": "SUCCESS",
                  "data": {
                    "id": "e1104336-d582-4779-b75b-64deb53e9f6c",
                    "name": "Example Organization",
                    "createdAt": "2025-07-25T08:46:01.485Z",
                    "updatedAt": "2025-07-25T08:47:11.988Z"
                  },
                  "timestamp": "2025-07-25T08:51:35.317Z"
                }
              }
            }
          },
          "404": {
            "description": "Account not found for the given ID.",
            "content": {
              "application/json": {
                "example": {
                  "status": "FAILURE",
                  "message": "Account with id - e1104336-d582-4779-b75b-64deb53e9f6c does not exists.",
                  "errorCode": "FAILED_ACCOUNT_RETRIEVAL",
                  "timestamp": "2025-07-25T08:52:25.740Z",
                  "path": "/agent-backend/accounts/e1104336-d582-4779-b75b-64deb53e9f6c"
                }
              }
            }
          }
        },
        "tags": ["Account"]
      }
    },
...
  },
  "info": {
    "title": "Agent Backend Service",
    "description": "✨The Agent Backend Service is a scalable, modular backend system designed to power intelligent agent-based applications. It provides robust APIs for authentication, user management, document processing, conversation tracking, training resources, and more. Built with flexibility in mind, it integrates seamlessly with external services while ensuring security, extensibility, and high availability.\n\n# 🔐 Authentication\n\nThis API uses a secure cookie named `token` to authenticate requests.\n\n- You must log in to receive this cookie.\n- When using **Try It Out** in the docs, **you must manually set the `token` cookie** in your browser.\n- You can find it in dev tools → Application → Cookies → {current url} -> `token`.\n\n**Note**: Scalar does not automatically send cookies during \"Try it out\".\n",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "name": "prod",
      "url": "https://zeno.sarvaha.ai/agent-backend"
    },
    {
      "name": "dev",
      "url": "http://localhost:7082/agent-backend"
    }
  ],
  "components": {
    "schemas": {
...
      "AddOrUpdateAccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "987f6543-e21a-45bc-94d1-112233445566"
          },
          "name": {
            "type": "string",
            "example": "Acme Inc."
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
    }
  }
}